hapurdeal/requirements.txt
dj-database-url==0.5.0
Django==3.0.7
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-cors-headers==3.4.0
django-embed-video==1.4.3
django-heroku==0.3.1
django-import-export==2.7.1
django-jet==1.0.8
django-js-asset==2.0.0
django-memcached==0.1.2
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==9.0.1
psycopg2==2.8.5
python-memcached==1.59
six==1.14.0
virtualenv==20.14.1
django-autoslug
django-colorfield
